ci/env.sh: Log `CI_BASE_BRANCH` (#33744)
When debugging CI failures it helps when I can see all the variables set
by `env.sh`.

Order environment variables in the output to match the order they are
set a bit better.

`CI_BASE_BRANCH` was not set at all for unknown environment.  Does not
matter much, but it seems consistent to include it in the list next to
all the other variables.